Move Card to components folder
